Make block more reusable
